Logging delegate for FaultExceptions
The exception is ususally caught at the boundries of a WCF object to
massage a fault exception to pass through the WCF channel.  As such I
added the option of providing a logging delegate in case one wanted to
log the error its own side of the WCF boundry. So components can do
their own logging and other components can catch the Exception Fault and
log it on their end