First test of Exit State transition success
Added the pieces to the test implementation and ran it through. works.
The Exit state from a substate pushes back ExitState transition that the
super state intercepts and queries its own result transitions for that
id.  Similar to defer but there is nothing for the super state to do at
that point