Added defered message formaters for Log Methods
In this way, the formating is never invoked unless the log level is
sufficiently high to have the message logged.  For example, you could
have a String.Format method creating your log string. Unless you use the
defered method in the logger, this string formating will occur even if
the message is never to be logged